requirements.txt
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.2.4
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.2
gunicorn==19.9.0
httpie==1.0.3
idna==2.8
Markdown==3.1.1
Pillow==6.2.1
Pygments==2.4.2
pytz==2019.2
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.3
